requirements.txt
Django == 4.0.5
pillow

#Instalar django_interface para personalizar panel 
django-admin-interface

django-crispy-forms

django-rest-framework

social-auth-app-django